requirements.txt
Django==1.8.13
tornado==4.3
django-redis
djangorestframework==3.1.1
django-filter==0.13.0
django-excel-response3==1.51
MySQL-python==1.2.5
pillow

requests==2.10.0
pika==0.10.0
amqp==1.4.6
kombu==3.0.26
celery==3.1.18
django-celery==3.1.17
eventlet==0.19.0
xlrd==1.0.0
xlwt==1.1.2
ansible==1.9.6
